requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
attrs==25.3.0
bracketool==0.1
cfgv==3.4.0
diagrams==0.24.4
discord.py==2.5.2
distlib==0.3.9
filelock==3.18.0
frozenlist==1.7.0
graphviz==0.20.3
identify==2.6.12
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.4.4
nodeenv==1.9.1
platformdirs==4.3.8
pre_commit==4.2.0
propcache==0.3.2
PyYAML==6.0.2
virtualenv==20.31.2
yarl==1.20.1
